+++ python_modules/libraries/dagster-airflow/tox.ini
@@ -23,7 +23,7 @@ commands =
   !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster -e dagit'
   localdb-airflow1: airflow initdb
   localdb-airflow2: airflow db init
-  default: pytest -vv {posargs}
+  default: pytest -m "not requires_airflow_db" -m "not requires_local_db" -vv {posargs}
   localdb: pytest -m requires_local_db -vv {posargs}
   persistantdb: pytest -m requires_persistent_db -vv {posargs}
